fix(aq3): genuine windows evidence artifact, daemon-bootstrap lib.rs split, AC8 citation, orphan harness removed, readiness timeout
- ATM-QA-101: replace the hand-authored windows skipped_no_tmux evidence
  pair with the byte-for-byte harness output from CI run 33126990425
  (started_at_ns + note now present, matching write_evidence()).
- ARCH-001/RULE-003: extract compose_queue_workers, the Herdr doctor
  adapters, and build_replacement_handler into a new sibling module
  crates/atm-daemon-bootstrap/src/replacement_handler.rs, following the
  queue_drain.rs/owner_gate.rs pattern; lib.rs drops to 761 non-test
  lines (cap 1000). Updates the architecture guard that pins
  build_replacement_handler's location and the benchmark-daemon
  isolation guard's search boundary to match.
- ATM-QA-102: cite CI run 33128710126 at head 8b8071793 in AC8 (all
  three Test lanes green), retaining the prior citation for context.
- ATM-QA-104: delete the orphan scripts/phase-aq/run_aq3_idle_drain_sweep_evidence.py
  harness (no EVIDENCE_DIR_BY_SCRIPT entry, no tests) and its committed
  local transcripts; document that deliverable 3 (recovery sweep) is
  covered by queue_drain.rs's unit/integration tests and that
  run_aq3_tmux_idle_drain_evidence.py is AQ3's sole live harness.
- FTQ-001: replace run_aq3_tmux_idle_drain_evidence.py's blocking
  readline() daemon-readiness wait (which ignored --timeout when the
  daemon produced no output) with a reader-thread/queue drain that
  honors the deadline on every read; add unit tests for both the
  timeout and ready-marker paths.
